+++ config/dev.exs
@@ -1,10 +1,13 @@
 use Mix.Config
 
 config :ex_cldr,
-  default_locale: "en-001",
+  # default_locale: "en-001",
   # locales: ["root", "fr", "zh-Hant", "en-GB", "bs", "pl", "ru", "th", "he", "af"],
   locales: [],
   gettext: Cldr.Gettext,
   precompile_number_formats: ["¤¤#,##0.##"],
   precompile_transliterations: [{:latn, :arab}, {:arab, :thai}],
   json_library: Poison
+
+config :gettext,
+  default_locale: "en"